+++ src/custom/api/gnosisProtocol/api.ts
@@ -24,6 +24,9 @@ import MetadataError, {
   MetadataApiErrorObject,
 } from './errors/MetadataError'
 
+import { DEFAULT_NETWORK_FOR_LISTS } from 'constants/lists'
+import { GAS_FEE_ENDPOINTS } from 'constants/index'
+
 function getGnosisProtocolUrl(): Partial<Record<ChainId, string>> {
   if (isLocal || isDev || isPr || isBarn) {
     return {
@@ -304,6 +307,20 @@ export async function uploadAppDataDoc(params: UploadMetadataParams): Promise<vo
   }
 
   await response.json()
+}
+
+export interface GasFeeEndpointResponse {
+  lastUpdate: string
+  lowest: string
+  safeLow: string
+  standard: string
+  fast: string
+  fastest: string
+}
+
+export async function getGasPrices(chainId: ChainId = DEFAULT_NETWORK_FOR_LISTS): Promise<GasFeeEndpointResponse> {
+  const response = await fetch(GAS_FEE_ENDPOINTS[chainId])
+  return response.json()
 }
 
 // Register some globals for convenience
