+++ src/custom/api/gnosisProtocol/mock.ts
@@ -1,7 +1,7 @@
 import { ChainId } from 'state/lists/actions'
 import deterministicHash from 'utils/deterministicHash'
 import { MetadataKind } from 'utils/metadata'
-import { AppMetadata, UploadMetadataParams } from './gnosisProtocolApi'
+import { AppMetadata, UploadMetadataParams } from './api'
 
 const appDataDoc = {
   version: '1.0.0',
